separate SLO route functions into different files (#208261)
Resolves [#207879](https://github.com/elastic/kibana/issues/207879)

This PR separates the different SLO route functions into separate files,
each with a single responsibility. The functions are still exported via
a single function, `getSloRouteRepository`.

---------

(cherry picked from commit a133dc57a7c2e95a080c0026b143472a35b69a1a)
