+++ x-pack/solutions/observability/plugins/slo/server/routes/slo/disable_slo.ts
@@ -1,0 +1,66 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { manageSLOParamsSchema } from '@kbn/slo-schema';
+import { executeWithErrorHandler } from '../../errors';
+import {
+  DefaultSummaryTransformManager,
+  DefaultTransformManager,
+  KibanaSavedObjectsSLORepository,
+} from '../../services';
+import { ManageSLO } from '../../services/manage_slo';
+import { DefaultSummaryTransformGenerator } from '../../services/summary_transform_generator/summary_transform_generator';
+import { createTransformGenerators } from '../../services/transform_generators';
+import { createSloServerRoute } from '../create_slo_server_route';
+import { assertPlatinumLicense } from './utils/assert_platinum_license';
+import { getSpaceId } from './utils/get_space_id';
+
+export const disableSLORoute = createSloServerRoute({
+  endpoint: 'POST /api/observability/slos/{id}/disable 2023-10-31',
+  options: { access: 'public' },
+  security: {
+    authz: {
+      requiredPrivileges: ['slo_write'],
+    },
+  },
+  params: manageSLOParamsSchema,
+  handler: async ({ response, request, context, params, logger, plugins }) => {
+    await assertPlatinumLicense(plugins);
+
+    const spaceId = await getSpaceId(plugins, request);
+    const dataViews = await plugins.dataViews.start();
+
+    const sloContext = await context.slo;
+    const core = await context.core;
+    const scopedClusterClient = core.elasticsearch.client;
+    const soClient = core.savedObjects.client;
+    const esClient = core.elasticsearch.client.asCurrentUser;
+    const dataViewsService = await dataViews.dataViewsServiceFactory(soClient, esClient);
+    const repository = new KibanaSavedObjectsSLORepository(soClient, logger);
+
+    const transformGenerators = createTransformGenerators(
+      spaceId,
+      dataViewsService,
+      sloContext.isServerless
+    );
+    const transformManager = new DefaultTransformManager(
+      transformGenerators,
+      scopedClusterClient,
+      logger
+    );
+    const summaryTransformManager = new DefaultSummaryTransformManager(
+      new DefaultSummaryTransformGenerator(),
+      scopedClusterClient,
+      logger
+    );
+
+    const manageSLO = new ManageSLO(repository, transformManager, summaryTransformManager);
+
+    await executeWithErrorHandler(() => manageSLO.disable(params.path.id));
+    return response.noContent();
+  },
+});
