+++ x-pack/solutions/observability/plugins/slo/server/routes/slo/get_suggestions.ts
@@ -1,0 +1,28 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { executeWithErrorHandler } from '../../errors';
+import { GetSLOSuggestions } from '../../services/get_slo_suggestions';
+import { createSloServerRoute } from '../create_slo_server_route';
+import { assertPlatinumLicense } from './utils/assert_platinum_license';
+
+export const getSLOSuggestionsRoute = createSloServerRoute({
+  endpoint: 'GET /internal/observability/slos/suggestions',
+  options: { access: 'internal' },
+  security: {
+    authz: {
+      requiredPrivileges: ['slo_read'],
+    },
+  },
+  handler: async ({ context, plugins }) => {
+    await assertPlatinumLicense(plugins);
+
+    const soClient = (await context.core).savedObjects.client;
+    const getSLOSuggestions = new GetSLOSuggestions(soClient);
+    return await executeWithErrorHandler(() => getSLOSuggestions.execute());
+  },
+});
